# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/track-groups": {
      "get": {
        "operationId": "TrackGroupController_list",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Which axis to list",
            "schema": {
              "enum": [
                "DECADE",
                "GENRE",
                "FAME",
                "CHART",
                "SPECIAL",
                "ARTIST"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackGroupDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Curated sets of songs anyone can play, library or not",
        "tags": [
          "Api"
        ]
      }
    },
...
    "/track-groups/{slug}": {
      "get": {
        "operationId": "TrackGroupController_bySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackGroupDto"
                }
              }
            }
          },
          "404": {
            "description": "No such group, or not for you"
          }
        },
        "summary": "One group, by the name in its URL",
        "tags": [
          "Api"
        ]
      }
    },
...
    "/me/status": {
      "get": {
        "operationId": "MeStatusController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Streak, daily and speed run state, at once",
        "tags": [
          "Api"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sopranino API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "unpaused_session"
      }
    },
    "schemas": {
...
      "StreakStatusDto": {
        "type": "object",
        "properties": {
          "currentStreak": {
            "type": "number",
            "description": "Current daily streak (consecutive win days)"
          },
          "bestStreak": {
            "type": "number",
            "description": "Best daily streak ever"
          },
          "playedToday": {
            "type": "boolean",
            "description": "Whether the user has won today"
          },
          "streakAtRisk": {
            "type": "boolean",
            "description": "Whether a gap exists and streak could be lost"
          },
          "canSaveStreak": {
            "type": "boolean",
            "description": "Whether the user can save their streak with freezes"
          },
          "gapDays": {
            "type": "number",
            "description": "Number of missed days in the gap"
          },
          "freezesAvailable": {
            "type": "number",
            "description": "Number of freezes available"
          },
          "freezeCost": {
            "type": "number",
            "description": "Number of freezes needed to bridge the gap"
          },
          "isTrusted": {
            "type": "boolean",
            "description": "Whether the user is trusted (can earn freezes)"
          }
        },
        "required": [
          "currentStreak",
          "bestStreak",
          "playedToday",
          "streakAtRisk",
          "canSaveStreak",
          "gapDays",
          "freezesAvailable",
          "freezeCost",
          "isTrusted"
        ]
      },
...
      "TrackGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable id, used to start a round"
          },
          "type": {
            "type": "string",
            "enum": [
              "DECADE",
              "GENRE",
              "FAME",
              "CHART",
              "SPECIAL",
              "ARTIST",
              "IMPORTED"
            ],
            "description": "Which axis this groups by"
          },
          "name": {
            "type": "string",
            "example": "1980s",
            "description": "What the tile is called"
          },
          "slug": {
            "type": "string",
            "example": "1980s",
            "description": "In the URL, so a shared link survives a rename"
          },
          "trackCount": {
            "type": "number",
            "example": 490
          },
          "imageUrl": {
            "type": "string",
            "description": "Album art of the best known song in the group, absent if none of them has any"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "trackCount"
        ]
      },
...
      "MeStatusDto": {
        "type": "object",
        "properties": {
          "streak": {
            "description": "Its playedToday means won today: the streak counts wins",
            "allOf": [
              {
                "$ref": "#/components/schemas/StreakStatusDto"
              }
            ]
          },
          "dailyPlayedToday": {
            "type": "boolean",
            "description": "Whether today's daily was finished, won or lost"
          },
          "speedRunBest": {
            "type": "number",
            "example": 42,
            "description": "Best speed run score, 0 if none"
          }
        },
        "required": [
          "streak",
          "dailyPlayedToday",
          "speedRunBest"
        ]
      },
...
    }
  }
}